workflow: recover from canceled workflows
This work focuses on cancelation management when it is caused by a session
worker quitting or becoming unavailable. The workflow section that uses the
session attempts to re-run the relevant activities.

Workflow cancelation (root context) has the same results, the workflow returns
and the defer is executed.
